more api separation
