+++ zeton/__init__.py
@@ -6,8 +6,7 @@ Aplikacja: system żetonowy ucznia/dziecka
 from flask import Flask
 import os
 
-from . import auth, views, db
-from zeton.api import api
+from zeton import api, auth, views, db
 from zeton.custom_jinja2_filters import jinja2_ban_datetime_filter
