+++ zeton/api/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('api', __name__, url_prefix='/api')
+
+from . import bans, child
